解决了聚合根 RowVersion 映射问题；解决了值对象不能更新问题。
