+++ src/ServerApi/Adnc.Core.Shared/Entities/Domain/AggregateRoot.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using Adnc.Core.Shared.Entities;
+﻿using Adnc.Core.Shared.Entities;
+using System.ComponentModel.DataAnnotations;
 
 namespace Adnc.Core.Shared.Domain.Entities
 {
     public abstract class AggregateRoot : Entity, IAggregateRoot<long>, IConcurrency
     {
-        /// <summary>
-        /// 并发控制列
-        /// </summary>
-        public DateTime? RowVersion { get; set; }
+        public byte[] RowVersion { get; set; }
     }
 }
