+++ src/ServerApi/Portal/Adnc.Ord/Adnc.Ord.Migrations/Adnc.Ord.Migrations.csproj
@@ -16,6 +16,11 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Migrations\20210302003516_Update20210302-02.cs" />
+    <Compile Remove="Migrations\20210302003516_Update20210302-02.Designer.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\..\..\Adnc.Infr.EfCore\22.Adnc.Infr.EfCore.csproj" />
   </ItemGroup>
